+++ misc.h
@@ -13,7 +13,7 @@
 #define ELEMS_ON_SCREEN 12
 #define DEBUG 1
 
-typedef std::pair<double, double> val_and_base;
+typedef std::pair<int, int> val_and_base;
 typedef const std::vector<val_and_base> & vals_and_bases;
 typedef std::pair<val_and_base, double> value_and_probability;
 typedef std::vector<std::pair<std::vector<val_and_base>, double>> complex_distribution;
@@ -22,6 +22,6 @@ void al_assert (bool check, std::string message);
 int toi (size_t src);
 unsigned int tou (int src);
 int fuzzycmp (double a, double b = 0.0);
-double sum (std::vector<double> values);
-double min (std::vector<double> values);
-double max (std::vector<double> values);
+int sum (std::vector<int> values);
+int min (std::vector<int> values);
+int max (std::vector<int> values);
